fc comm using grouped send/recv
fix

fix

fix

fix
